Hide non-current toc lists in foldable tocs
Removing the default hiding of the sub lists lead to the collapsible functionality
not working. This was because active/current levels are shown by matter of being
child of a current class.

This commit reverses the logic and hides the sublists only when the parent list is active
but the list item itself is not current.