Hide non-current sublevels in site toc only
Before, the rules for the page toc and the site toc where mixed.
This made it hard to find a way that consistently hides the non-current elements
in the site toc, but always shows all elements in the page toc.

Now a wrapper is required around to the site toc to hide the lower levels.