Fix: Page toc levels (#82)
* Show all levels of page content toc

This removes the one style rule that was hiding the lower nagivation elements.

* Hide non-current toc lists in foldable tocs

Removing the default hiding of the sub lists lead to the collapsible functionality
not working. This was because active/current levels are shown by matter of being
child of a current class.

This commit reverses the logic and hides the sublists only when the parent list is active
but the list item itself is not current.

* Add example page for nested pages

This is helpful to test the toc nesting deeper than only one level.

* Add size scale and reduce toc indent

* Fix toc breakpoints

* Remove unused styles in toc component

* Hide non-current sublevels in site toc only

Before, the rules for the page toc and the site toc where mixed.
This made it hard to find a way that consistently hides the non-current elements
in the site toc, but always shows all elements in the page toc.

Now a wrapper is required around to the site toc to hide the lower levels.