Explicitly define col widths
The automatic with `col` did not work quite right on pages generated from RST.
I am not sure why, but defining the col widths explicitly solves the issue.

For what it's worth, I also think it looks better and seems to behave nicer
and more predictable on other pages too.