Remove and comment unused styles
Some styles I have remove that related to the difference between the sidebar
picker and the hovering badge picker.

I have now styled the picker to be always the badge style, so the special
styles are not needed.

I also commented a bunch of styles that I did not see any use of. Maybe
@stormheg can say more. He had created the initial file I was workin off.