Fix version picker (#90)
* --wip-- [skip ci]

* Remove apidoc generation

We don't need that and it just keeps generating useless files in the repo.

* Add Wagtail docs version picker to layout override

Extends the previous idea of just copying in the HTML that is injected by read
the docs.

This also changes to pull the css from the readthedocs online source.

* Add version picker overrides back

* Create little batch version of the picker

* Remove and comment unused styles

Some styles I have remove that related to the difference between the sidebar
picker and the hovering badge picker.

I have now styled the picker to be always the badge style, so the special
styles are not needed.

I also commented a bunch of styles that I did not see any use of. Maybe
@stormheg can say more. He had created the initial file I was workin off.

* Fix previous commit

I think I removed to much on the last one. This adds a few styles back.

Co-authored-by: Tibor Leupold <tibor.leupold@torchbox.com>