Reduce colors for link style
The theme is already using quite a few colors. In an attempt to
consolidate all these colors and reduce the number of used colors, the
teal on hover is now replaced with the underline/border color (green).

On hover and active, line and text are now of the same color. This feels
a bit cleaner to me.