Link hover style (#152)
* added teal for hover and active state

* added color for active state

* fixed error

* Remove visited link style

All links in a documentation will be visited very soon. The value of
coloring them different (especially the same as the hover style) is
therefore questionable.

* Reduce colors for link style

The theme is already using quite a few colors. In an attempt to
consolidate all these colors and reduce the number of used colors, the
teal on hover is now replaced with the underline/border color (green).

On hover and active, line and text are now of the same color. This feels
a bit cleaner to me.

* Use underline style insread of border

The border was too far way and is influenced by the padding. This is not
really what we wanted.

* Remove underline from logo link

* Add underline to first level main toc links again

* Close underline gap in pager links

* Remove underline from source buttons

* Remove underline from footer links

Co-authored-by: PaarthAgarwal <86092410+PaarthAgarwal@users.noreply.github.com>