Add theme requirement to theme's docs requirements
Because with myst_parser we now have actual depenencies that the theme
needs, we need to make sure these depenendies are installed when we are
building the docs of the theme with the theme (so meta...).

Previously we might have gotten away without that, because most
dependencies have been installed by Readthedocs anyhow. But,
`myst_parser` is not installed by RTD. So now we really need to install
the theme requirements to host the themes docs with the latest version
of the theme itself.